+++ xcloud-components-core/src/main/java/com/wl4g/components/core/bean/ci/PipeStepBuilding.java
@@ -22,25 +22,21 @@ import com.wl4g.components.core.bean.BaseBean;
 public class PipeStepBuilding extends BaseBean {
 	private static final long serialVersionUID = 6815608076300843748L;
 
-	private Integer pipeId;
-
+	private Long pipeId;
 	private String preCommand;
-
 	private String postCommand;
-
 	private Integer refType;
 
 	// ==================Expand==================
 
 	private List<PipeStepBuildingProject> pipeStepBuildingProjects;
-
 	private String ref;
 
-	public Integer getPipeId() {
+	public Long getPipeId() {
 		return pipeId;
 	}
 
-	public void setPipeId(Integer pipeId) {
+	public void setPipeId(Long pipeId) {
 		this.pipeId = pipeId;
 	}
