+++ xcloud-components-core/src/main/java/com/wl4g/components/core/bean/ci/PipeStepInstanceCommand.java
@@ -20,17 +20,17 @@ import com.wl4g.components.core.bean.BaseBean;
 public class PipeStepInstanceCommand extends BaseBean {
 	private static final long serialVersionUID = 6815608076300843748L;
 
-	private Integer pipeId;
+	private Long pipeId;
 
 	private String preCommand;
 
 	private String postCommand;
 
-	public Integer getPipeId() {
+	public Long getPipeId() {
 		return pipeId;
 	}
 
-	public void setPipeId(Integer pipeId) {
+	public void setPipeId(Long pipeId) {
 		this.pipeId = pipeId;
 	}
